refactor(credential): move ScopedCredentialAccessor to engine
The engine-runtime accessor with key-based access scoping
sits at the engine layer where it actually wires up, alongside
EngineCredentialAccessor. The crate-internal stub
(NoopCredentialAccessor + default_credential_accessor) stays in
nebula-credential because CredentialContext consumes them in
the same crate and credential cannot depend on engine without
inverting the layer graph.

Tests for the moved type (including the MockAccessor helper)
travel with the impl to engine/credential/scoped_accessor.rs.
